separando las vistas de la aplicacion
