+++ apps/account/future_views/__init__.py
@@ -1,0 +1,8 @@
+from .signup import SignUp
+from .login import Login
+from .reset_password import ResetPassword
+from .downloads import Downloads
+from .ranking_players import RankingPlayers
+from .ranking_guilds import RankingGuilds
+from .request_token import RequestToken
+from .request_username import RequestUsername
