+++ apps/account/views.py
@@ -61,123 +61,16 @@ from django.utils.translation import ugettext as _
 # importando las configuracines de django
 from core import settings
 
-
-# clase usada para el registro de usuarios
-class Create(CreateView):
-    success_url = reverse_lazy('account:exito')
-    template_name = 'account/registro.html'
-    model = Account
-    form_class = CreateUserForm
-
-    def get(self, request):
-        lenguaje(request)
-        return super(Create, self).get(request)
-
-    def post(self, request):
-        lenguaje(request)
-        return super(Create, self).post(request)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        if 'refer' in self.request.GET:
-            context['refer_id'] = self.request.GET.get('refer')
-        return context
-
-    def form_valid(self, form):
-        key = aleatorio(40)
-        self.object = form.save(commit=False)
-        new_password = self.object.micryp(self.object.password)
-        self.object.password = new_password
-        self.object.address = key
-        if 'refer_id' in self.request.POST:
-            self.object.refer_id = self.request.POST['refer_id']
-        self.object.save()
-        try:
-            send_mail(
-                _('Bienvenido a ') + settings.SERVERNAME,
-                'Content',
-                settings.EMAIL_HOST_USER,
-                [self.object.email],
-                html_message=get_mail_register(self.object.login, key),
-                )
-        except Exception as err:
-            raise SendMailException(err)
-        return HttpResponseRedirect(self.get_success_url())
-
-
-# funcion usada para el login
-# Refactorizando la funcion de login
-class Login(View):
-
-    template_name = 'account/login.html'
-    template_name_login = 'account/logon.html'
-    modelA = Account
-    modelB = Top
-    form = CustomLoginForm
-
-    def get(self, request):
-        lenguaje(request)
-        form = self.form()
-
-        if 'g1jwvO' in request.session:
-            userinfo = self.modelA.objects.get(id=request.session['g1jwvO'])
-            pjinfo = self.modelB.objects.filter(account_id=userinfo.id)
-            context = {
-                'session': userinfo,
-                'personajes': pjinfo
-            }
-            return render(request, self.template_name_login, context)
-        else:
-            pass
-
-        context = {
-            'form': form
-        }
-        return render(request, self.template_name, context)
-
-    def post(self, request):
-        lenguaje(request)
-        form = self.form(request.POST or None)
-
-        if form.is_valid():
-            try:
-                a = self.modelA.objects.get(login=form.cleaned_data['login'])
-            except Account.DoesNotExist:
-                context = {
-                    'key': _('Nombre de usuario o password incorrecto'),
-                    'form': form
-                }
-                return render(request, self.template_name, context)
-
-            b = a.micryp(form.cleaned_data['password'])
-            # Validando que las contraseñas coincidan
-            if a.password == b:
-                # Validando que la cuenta no este baneada
-                if a.status == 'OK':
-                    request.session['g1jwvO'] = a.id
-                    if 'g1jwvO' in request.session:
-                        return redirect('account:login')
-                else:
-                    context = {
-                        'key': _('Tu cuenta esta baneada'),
-                        'form': form
-                    }
-                    return render(request, self.template_name, context)
-
-            else:
-                context = {
-                    'key': _('Nombre de usuario o password incorrecto'),
-                    'form': form
-                }
-                return render(request, self.template_name, context)
-
-        else:
-            context = {
-                'key': _('Rellene todos los campos correctamente'),
-                'form': form
-            }
-            return render(request, self.template_name, context)
-
+from .future_views import (
+    SignUp, 
+    Login, 
+    ResetPassword, 
+    Downloads,
+    RankingPlayers,
+    RankingGuilds,
+    RequestToken,
+    RequestUsername
+) 
 
 # funcion usada para cerra session
 def logout(request):
@@ -256,44 +149,6 @@ class ClaseExito(View):
         lenguaje(request)
         # Se renderiza el template con el contexto.
         return render(request, self.template_name)
-
-
-# Clase usada para la pagina de descargas.
-class Descargas(View):
-    model = Descarga
-    template_name = 'account/download.html'
-
-    def get(self, request):
-        lenguaje(request)
-
-        context = {
-            'descarga': self.model.objects.publicado(),
-        }
-        return render(request, self.template_name, context)
-
-
-# clase usada para la pagina del ranking del juego y paginarlo
-class ClasificacionPersonajes(ListView):
-    model = Top
-    template_name = 'account/top100.html'
-    queryset = player_top()
-    paginate_by = 20
-
-    def get(self, request):
-        lenguaje(request)
-        return super(ClasificacionPersonajes, self).get(request)
-
-
-# Clase usada para renderizar el top del juego y paginarlo
-class ClasificacionGremios(ListView):
-    model = Guild
-    template_name = 'account/top_g.html'
-    queryset = guild_top()
-    paginate_by = 20
-
-    def get(self, request):
-        lenguaje(request)
-        return super(ClasificacionGremios, self).get(request)
 
 
 # Funcion usada para recuperar password por correo
@@ -507,124 +362,3 @@ def desbuguear(request):
         return redirect('account:login')
 
 
-# Clase usada para re enviar el correo de activacion, en caso de que no llegue.
-class RequestToken(View):
-
-    template_name = 'account/envio_token.html'
-    model = Account
-    form = ResPassword
-
-    def get(self, request):
-        lenguaje(request)
-        form = self.form(request.POST or None)
-        context = {
-            'form': form
-        }
-        return render(request, self.template_name, context)
-
-    def post(self, request):
-        lenguaje(request)
-        form = self.form(request.POST or None)
-        context = {
-            'form': form
-        }
-        if form.is_valid():
-            try:
-                a = self.model.objects.get(login=form.cleaned_data['login'])
-            except Account.DoesNotExist:
-                context.update({
-                    'key': _('Cuenta no existe.')
-                })
-                return render(request, self.template_name, context)
-
-            b = datetime.datetime(2009, 1, 1, 0, 0,)
-            if a.availdt.year == b.year and a.availdt.month == b.month:
-                context.update({
-                    'key': _('Tu cuenta ya esta activada'),
-                })
-                return render(request, self.template_name, context)
-
-            if a.status != 'OK':
-                context.update({
-                    'key': _('Tu cuenta esta baneada')
-                })
-                return render(request, self.template_name, context)
-
-            if form.cleaned_data['email'] == a.email:
-                key = aleatorio(40)
-                a.address = key
-                a.save()
-                try:
-                    send_mail(
-                        _('Activacion de cuentas ')+settings.SERVERNAME,
-                        'Content',
-                        settings.EMAIL_HOST_USER,
-                        [a.email],
-                        html_message=get_mail_register(a.login, key)
-                    )
-                except SendMailException:
-                    context.update({
-                        'key': _('Error enviando correo al usuario')
-                    })
-                    return render(request, self.template_name, context)
-
-                context.update({
-                    'key': _('Se ha enviado el codigo de activacion al email')
-                })
-                return render(request, self.template_name, context)
-            else:
-                context.update({
-                    'key': _('El email no coincide con el usuario')
-                })
-                return render(request, self.template_name, context)
-        else:
-            context.update({
-                'key': _('Por favor rellena todos los campos correctamente.')
-            })
-            return render(request, self.template_name, context)
-
-
-class RequestUsername(View):
-    template_name = 'account/usernames.html'
-    model = Account
-    form = FormRequestPassword
-
-    def get(self, request):
-        lenguaje(request)
-        form = self.form(request.POST or None)
-        context = {
-            'form': form
-        }
-        return render(request, self.template_name, context)
-
-    def post(self, request):
-        lenguaje(request)
-        form = self.form(request.POST or None)
-        context = {
-            'form': form
-        }
-        if form.is_valid():
-            accounts = Account.objects.filter(email=form.cleaned_data['email'])
-            if len(accounts) == 0:
-                context.update({
-                    'key': _('No se encontraron cuentas asociadas al email')
-                })
-                return render(request, self.template_name, context)
-            try:
-                send_mail(
-                    _('Cuentas asociadas al correo ') + settings.SERVERNAME,
-                    'content',
-                    settings.EMAIL_HOST_USER,
-                    [form.cleaned_data['email']],
-                    html_message=get_mail_username(accounts)
-                )
-            except SendMailException as err:
-                context.update({'err': err})
-                return render(request, self.template_name, context)
-            context.update({
-                'key': _('Correo enviado exitosamente')
-            })
-            return render(request, self.template_name, context)
-
-        else:
-            return render(request, self.template_name, context)
